feat(ET-202): Retrieve and Decrypt message bodies.
Store message body metadata and extend the attachment table to include
the final pieces which are only available in the `Message` type.

Message bodies will be stored outside the database in the device's local
file system. A new parameter has been added to `MailContext` to control
the cache location for the messages.

Finally the handling of multi-part messages is delegated to future work
(ET-263).