fix(ET-223): Store conversation remote id on message
When the user has message view mode enabled it is possible that the
message conversation has not been downloaded yet. To ensure all
optimistic calculations proceed as expected we also store the
conversation remote id on the message.

This will be used by the message actions which will come in a future
patch.