feat(ET-815): Strip UTM parameters from HTML content
This patch adds the `proton-mail-html-transformer` crate which serves as
the foundation for all upcoming HTML content transformation passes
(proxy images, sanitization, etc...).

Currently the only pass available is the stripping of UTM trackers. This
feature was previously available as as standalone crate, but has
since been merged in this one. The features are still accessible as part
of the `utm` module.