Tidied up migrations and unnecessary transactions
Much of this is temporary, as the migration system is about to be
rewritten.

The transactions removed are those that were following the previous
approach whereby there was always a "tx" context, and not necessarily
needing an actual transaction. To avoid confusion, these have been
removed, and if needed, they can be thoughtfully re-added as may be
appropriate later.