refactor: Restore Authentication Store
This functionality was removed in the previous refactor and is required
so we can persist sessions to disk and execute our integration tests.

This patch restores a simplified version of was there before and
separates the in memory caching requirement from the actual storage
medium.

Finally, this patch also removes the `Auth` data from the `Session` type
as it can easily access the one present in `Proton` api.