chore: Replaced integration_tests feature with debug_assertions
The integration_tests feature was introduced as a way to use conditional
compilation for certain things needed by integration tests, but not
production code, such as applying Serialize to API response data. This
is needed because integration tests do not run in test mode, and so just
detecting test mode will not work.

A better way has been found, which is to detect the presence of the
debug_assertions flag, which should only be part of the dev profile.
This will therefore compile in the things needed for integration tests
in dev mode, but not in release mode.