fix: Corrected use of rusqlite in action-queue crate
As nothing in the ecosystem uses rusqlite directly, and the only use of
database functionality is through the stash crate, all crates using the
database should use the exported stash types rather than rusqlite ones.
This is different to a situation where types from a crate are widely
used, in which case re-exporting is not desirable.