test: Restore attachment tests
This patch restores the attachment test to working order as well as
previous behavior in regards to how the attachment data is store in the
database.

Additionally, neither `Conversation` or `Message` need the full
attachment information stored on the types as it is not possible to
correctly represent them until all the metadata is downloaded.

The full `Attachment` type is also removed from uniffi as
any operation that involves an attachment is transparently handled
by the common crate.

Finally, there is one missing part related to message body metadata that
will be resolved in a future patch.