fix(ET-1006): Restore contextual information for conversations
This patch adds the `ContextualConversation` type which represents a
a `Conversation` being opened in the context of a `Label`.

Contextual information for `Conversation` is obtain by superimposing the
`ConversationLabel` for the active label over the type. This in turn
restores the missing `time` attribute, which is only available on
the `ConversationLabel`.

Finally this patch replaces the `Conversation` type in
`proton-mail-uniffi` to be constructed from the `ContextualConversation`
as it can't be used correctly otherwise.