fix: Async functions need to run on our own runtime
Uniffi runs the async functions in async-std. There is a way to use
tokio but it relies on async-compact. If it does not detect a runtime
associated with the thread, it will instanciate a thread local tokio
runtime. Unfortunately this is not compatible with our code as we
require the multi-threaded version.

To work around this all public exported async functions need to run a
task on a runtime we manage with `uniffi_async`. The runtime is
initialized on first use. There may be another way to manage this, but
it requires further investigation.