fix(ET-1049): Replace (name, email) column pair with single name_or_addr
When signing in, the API accepts an account's username as well as any
of its addresses. This caused some ambiguity in what we were storing
in the DB. This change replaces the non-null `email` column and
nullable `name` column in the DB and associated session types with a
single `name_or_addr` column.