fix(ET-1074): Tether leak
The previous change to stash remove the garbage collection in favor of a
more aggressive approach where workers are pruned as soon as the
connection closes.

It was then observed that the `Drop` implementation for `Tether` is
prone to leaking connections. Manually checking the strong
count of the handle can lead to incorrect results. It is possible for
the handle to still be in use by commands when the last `Tether`
instance drops.

We address this by wrapping the inner data of the tether in an Arc
that safely and correctly the close connection command when all instances
of that `Tether` are dropped. Its was internally wrapped to avoid
causing a giant set of changes where all `Tether` instances need to be
updated to `Arc<Tether>`.