fix(ET-1214): Do not trigger callbacks for recenlty synced elements
When syncing new pages to the database do not trigger notifications for
insert or updates for these new elements as they are redundant.

The redundancy comes from the fact that the fetching next pages is performed
directly by the user and we already return the new page at the same
time.

For this to work correctly the `DataSource` now returns the list of
inserted elements so we can correctly track these elements.