refactor: Add execution contexts to the action queue
Every action can now have access to an execution context which
is retrieved from the queue when the actions are executed.

Contexts need to be registered ahead of time so they can work correctly
with queued actions which are serialized.

It would have been easier to have the context defined as trait on the
queue, but that would not allow us to have one queue which is shared
between core and mail contexts. This is desirable later when we
have actions that operate on core types and follow up mail actions which
depend on the execution of the core actions.

Finally the expanded tests was removed as it is no longer compatible
with the new queue interface. Testing coverage is handled by
the side bar tests.