refactor: Make queued action runnable in multithread async runtime
The previous implementation could not run the `QueuedAction` in
multi-threaded runtime environment as the futures were not sendable.

Changes have been made to make sure the `Handler` and the
`Action::RemoteOutput` are sendable.

Some functions on the `Queue` have been generalized so they do not
need to hold onto a reference to `self`.