refactor: Remove Session parameter from `Handler::apply_remote`
It was observed that now that we are passing the context to the action,
we can pass in any type of data necessary.

Most actions in core and mail will need access to their respective
context at some point so we can easily extract the session from those
instances.

Finally, the `UserContext` creation function now return the instance
wrapped in `Arc` so it can be added to the queue context. This change
is also desirable for the future when we create other BU contexts that
build over the same core use context.