refactor(ET-1863): Extract counters from labels
# Context

Labels are supposed to be BU independent.
While we had them so far in Mail, we also want to reuse the same
datastructure in contact list.

However, in our codebase, we made Labels mail-dependent, by including
in them number of messages and number of conversations.

This commit is first of two that extract those fields into separate
tables. It is focusing only on Message counters, leaving Conversation
counters for the future work.