refactor(ET-1894): Draft embedded attachments
Refactor the `Draft` type to hold onto the decrypted body so we
can also load the inline and pgp embedded attachments when drafting
a reply.

We also extend `DecryptedMessageBody` to optionally skip pre-fetching,
this is useful for testing and other scenarios where this is guaranteed
(e.g.: Draft Reply).

Finally we change some result types from `MailboxResult` to
`MailContextResult` for consistency.