fix: Do not override draft save state from server
If for some reason the are more than one draft save action, it is
possible that the local state gets overridden by the server
response leading to some odd behavior where the message info
can change in between action execution.

To prevent this we only update information from the server which
we do not manually override.