refactor: Cleanup the fix for release
The change that introduced the cancellable task did not correctly
hook up the custom runtime instance to the spawned task.

We addressed this issue by duplicating the code to ship this fix
quickly.

This change cleans up the code by using the `TaskSpawner` trait to
abstract the spawning of the task. Each `Context` and `UserContext` now
has a `spawn_with` method where alternate spawn implementations can be
specified.