refactor(ET-2154): Image proxy transformation
Other than providing a method to undo the proxy transformation (required
for drafts) we also make sure that we attempt to proxy other sources
which can potentially point to images.

This patch also ensures that srcset elements are correctly proxied
and `data:` urls are ignored as they can't be proxied.

Finally, we moved the proxy methods to the `remote_content`
module.