feat(ET-404): Use mail session
# Context

iOS receives the device token before first login, whenever user
accepts push notifications.

That means, we need to store it in the account DB and be able to
store and retrieve it using MailSession, instead of MailUserSession.

However, API call needs user ID, so that there is a need for three,
not two actions:

* Retrieving the token from DB (if user logs in into second account,
  we need the same device token)
* Storing the token from DB (both before logging in and after)
* Registering the token in our API (only after logging in)