feat(ET-480): Attachment Upload API
Introduces the necessary machinery to encrypt and upload the attachment
to the proton servers.

Unfortunately muon does not have multi-part form data support. Due to
time constraints we generate this data locally and will look into
patching muon when time permits.

Finally the `send_mail` example has been updated to test the attachment
upload as we need an existing message to do so. While we currently can't
send messages, this will be addressed with the draft attachment upload
feature.