feat(ET-1407): Add new attachment to drafts
Introduce `Draft::add_attachment` and the `AttachmentUpload` action the
user can add new attachment to drafts.

Existing updates to draft messages via the `Save` actions had to be
further reduced so that they do not overwrite local state setup by
chained actions. This also applies to AttachmentMetadata updates
from Conversations and messages.

The tracking of the last save action now uses the recorded ids in the
database rather than the manual book keeping from before.

The dependency on self error was removed from queue_or_replace as we can
actually need to be able to do link replacement actions for drafts to
avoid running duplicate logic if the replacement fails. This is safe
to do as we still remove the self dependency before save.