fix(ET-1470): Make sure draft save saves the correct state
This issue went unnoticed due to the lack of concurrent queue executors.
When saving the draft we want to make sure we only push to the server
the changes at the time the action was recorded.

There were some subtle race condition that could trigger from time to
time when re-reading the data from the database. With each save action
we now record the full state and we only refresh the remote ids before
upload.

Send is not as affected by this, since there can not be any
modifications to the message after the send action is queued.