feat(ET-2141): Subscribe to network even only once and reuse the same subscription in one executor
# Context

In previous MR we have introduced `WaitForOnline` trait.
That trait had `&self` reference, and the instance of it was shared
between all executors.

As a side effect of this design, every time there was a network failure,
we were subscribing to the event **again and again** instead of reusing subscribtion.

# Solution

This MR separates concerns - The shared instance is still immutable, but
now it is a "factory" - subscribing to the events and returning receiver.

The queue executor subscribes **once** and keeps mutable reference on a stack.

# Additional change

I also decided to introduce DummyWaitForOnlineSubscribion directly in the
action queue crate. That removes several identical implementations across
different crates and modules.