fix(ET-2316): Forwarding and replying with attachments
Update the new draft save to correctly work with attachments that
are inherited from a draft reply/forward. This stopped working as the
API automatically assigns new ids to these attachments and we need
to reflect this in on our internal state.

With this patch, the drafts now track their own attachment state so
there is no interference from external updates. When the draft is
created on the server we check and update all inherited attachments.

To make sure this works for external addresses (non-proton) we also
need to duplicate the attachment data so it is available when sending.