feat(ET-2507): Remove attachments from draft
Add support to remove attachments added to a draft. If we own the
attachment we need to delete from the server as well as locally. The
latter is necessary as there is no server event to delete attachments.

Finally we also mark the metadata as deleted to hide it from the
active attachment list.