chore(ET-2558): Initialize on context creation
# Context

In the last MR we have introduced the idempotent initialization.
This is needed for the sake of having legacy migration in the offline
mode.

This commit, separates initialization (that no longer exists as Uniffi
call). Instead, initialization happens automatically whenever we acquire
a new user context.

# Side effects

Acquiring a new user context may fail because of network issue. This
is something that our mobile devs will have to solve on their side.

# Unit tests

Bunch of our unit tests (120+) was depending on particular context
creation order. It is because, before initialization, we have to setup
mocks for our API.

This commit solves that issue in two ways:
* Whenever possible reorders the code so that user context is acquired
AFTER mocks are ready,
* Or as a last resort, acquires **uninitialized** context with manual
deferred initialization.

Note, the second solution is **only** for tests.