feat(ET-2592): Task Service
Introduce the `TaskService` which provides a way to spawn futures that
can be paused and resumed as required. This is necessary for iOS to
ensure that we don't have any background tasks that are running or using
resources which can cause the OS to kill the process.

We also provide an `UnpausableFuture` which, when used in a task spawned
via `TaskService` can be used to overwrite the paused state. This is
important for futures that hold onto critical resource that must
terminate. (E.g.: Sqlite Transaction).

This set the foundation for the next set of changes to ensure that we
can pause all work as the application enters the background and any open
transactions complete to release the db file locks.