refactor(ET-2592): Update background execution API
Move all logic to mail-common and update the execution of the background
work to return a status.

The uniffi side has also been updated with a default timeout and an
option to manually specify a desired timeout.

`BackgroundExecutionStatus::Executed` is currently not returned as this
requires terminating the executors when the queue is empty. This will
come in a separate MR.