fix(ET-2592): Replace r2d2 with basic replacement
It has been observed in crash reports that the default behavior of r2d2
is causing it to open database connections at very inopportune times,
leading to iOS terminating the application.

To address this we replace it with a very basic implementation that has
a more predictable behavior.

It is possible that setting r2d2's idle connection count to 0 could also
solve the issue, but we can't say with absolute certainty that is the
case. One the plus side this removes at least 3 thread workers per
database connection.