fix: Escape rendering of forwarded plain text messages
When we convert text messages to HTML, we inject some html elements to
preserve the spacing. If the HTML sanitizer find a sequence of `<>` it
would consider it an open HTML tag and inject a closing version at the
appropriate scope. This in turn leads to invalid html which then
stripped out later on.

To prevent this we escape all `>` and `<` occurrences in the text
document to avoid triggering this behavior.