fix: One transaction per event
On iOS we observed a grace period of about 200 ms on one device before
the process gets killed by the OS with an open transaction.

While we allow running transactions to finish, it has been noted that if
we have a long list of events, we can easily take more than 200ms to
apply them all.

To address this in the short term, we publish each event individually to
make sure the transactions are kept very short.