feat: Send tracing logs both to file and to iOS OS-Logger
This feature introduces `tracing_oslog` library that allows us
to send tracing logs and spans to iOS logger. It means, while debugging
the application, you do not need a separate log opened, as everything
is visible directly in XCode.

You can search, filter, clear directly in the IDE, instead of relying
on Console or terminal apps like `lnav`.

# File logging

File logging stays the same for the sake of testing and
sharing logs by Q&A. 

# Android

This commit focuses **only** on iOS.

# Double logs

Currently the iOS application sends all logs to Rust and to OS logger.
That creates duplicated entries in OS logger (but not in the file)
because with this commit the flow looks like this:

Swift -> OS Logger
    \
     +-> Rust -> File
           \
            +--> OS Loggger (duplicate)

It is, therefore, recommended to remove `logToUnifiedLoggingSystem` on
iOS side and rely only on Rust connection between `tracing`
and OS logger. Ideally it would look like this:

Swift -> Rust -> File
           \
            +--> OS Logger