feat: Enable Rollback Items
Add the Rollback action to flush the items which needs to be resynced
from the server when an action fails to execute.

The rollback was refactored to work with the action queue. The
itertools' chunks extension is not Send safe and we optimized the
request to fetch all the items in batches.

Message fetching has been updated to only fetch the metadata as the only
time we need to refetch the body is for drafts. The latter is handled
in the draft workflow.

Finally, the rollback action is queued every time we poll events.