fix(ET-2783): Ensure all logged in and intialized contexts are used
Fix `initialized_user_context_from_session` to instantiate a new
MailUserContext if none was ever created before. Previously this would
only work if the context was created at least once in the foreground.

Note that it is currently possible to have different contexts created
without the correct parameters for the session hook and status observer.
This will be addressed in a follow up.