refactor*: Rework Session Deletion Observation
Rather than specifying the subscriber on the `UserContext` every time,
have one subscription mechanism on the `Context` type itself.

Firstly, the `Context` is the current owner of the account db where all
the sessions live.

Secondly, it was possible to create new context that would bypass
subscription hooks that are required for everything to work correctly.

This new version ensures those hooks only need to be set up once per
`Context` instance and are valid for any `UserContext` that is created.