fix*: MailScroller on initialization may now decide to fetch next page in order to fill
when it has insufficient amount of items at hand, due to for example During clean install prefetcher downloads fraction of the page of inbox so it was possible to serve first page to the client which was less than requested making it looks like a bug. Now the scroller will decide if it needs to fill the page so on `fetch_more` it will serve correct amount of items.