feat*: ical: Improve error recovery
Most of our *.ics parsers are SAX-style, like:

```rust
// SOMEOBJECT;FOO=...;BAR=...;ZAR=...:VALUE

impl IcsRead<Property> for SomeObject {
    fn read(r: &mut IcsReader) -> Option<Self> {
        let mut foo = None;
        let mut bar = None;
        let mut zar = None;

        loop {
            let e = r.entry()?;

            if e.try_param(r, "FOO", &mut foo)
                || e.try_param(r, "BAR", &mut bar)
                || e.try_param(r, "ZAR", &mut zar)
            {
                continue;
            }

            if e.is_value() {
                break;
            }

            // throw "invalid parameter"
            e.burn(r);
        }

        Some(Self {
            foo,
            bar,
            zar,
            value: r.value()?,
        })
    }
}
```

This design seems to work overall great, but it's got an important shortcoming:
if `e.is_value()` never appears, this "sub-parser" will eat through all of the
remaining tokens ("entries"), never realizing it got itself into a corner.

As a specific example, if we try to parse an invalid-but-not-deadly-so:

```
SOMEOBJECT;FOO=""
UID:1234
```

... `SomeObject`'s parser will happily eat the newline and start going through
tokens that belong to `uid`.

This will _not_ cause us to return an invalid structure, because `r.entry()`
will report all errors anyway, it's just that next to the error reporting we
already have, we should jump out of a parser if it stumbles upon an unexpected
newline -- so that in cases like:

```
SOMEOBJECT;FOO=""
ORGANIZER;CN="someone":mailto:someone@localhost
```

... we report just one error (`unexpected newline`) instead of more (`unexpected
property: CN` etc.).