feat*: ical: Improve CLI, extract into own crate
Since now we're running the crate on production, I'm getting *.zips of
files that we're failing to parse (amazingly, in _all_ cases so far it's
because somebody doesn't stick to the standard, forcing us to implement
workarounds within the parser).

Anyway, analyzing those files by running the parser manually within Rust
tests is PITA, so I took a moment to implement a proper CLI tool for
working with *.ics files.

So far I think I'm going to be the only user, but since later we're
going to be refactoring the vcard crate as well, I imagine it becoming a
more useful tool in time (possibly with support for crypto operations).

But for the time being all I need is just a way of mass-analyzing *.ics
files.