fix: Correct cached status for conversations created by a draft
Ensure that the conversation created when composing a new message
is marked as having its messages so it can be view correctly offline
on mobile clients. Mobile clients always use the conversation viewer to
display their messages.

We also make sure that if the conversation has had their messages synced
at least once, we preserve this metadata on future updates.