fix: DraftSendResultWatcher not triggering after send
When operating in the sent only mode, we would not receive the
notification of the new record. This happened because the
`DraftMetadata` that we check for is deleted by the `Send` action.

It is safe to manually set this after the send action execution since
it is a direct result of a send action.