feat(ET-2568): Attach public key to messages
When the user has the mail setting enabled to attach public keys to the
their message we add the respective public key for the address to the
message if it does not yet exist.

Contrary to web, we will display the public key in the attachment list
at all times.