chore: Update lightningcss to fork of 1.0.0-alpha.66
I have decided to fork Lightningcss in order to bring support
for `-webkit-text-fill-color`. I realized that writing a workaround
in Proton would be very complex, while adding a support to the library
took two lines of code.

And since we crossed the Rubikon it allows us to fix all those small
QoL things like the lack of `Debug` or `Clone` in some structs etc.