fix(ET-3351): Out of bounds access in Draft::save
The list of attachments we attach to a draft save can only
contain attachments which have a remote id. Previously this filtering
needed to be done in 3 separate places, now we filter once when
we load the attachments.

The out of access crash was due to enumerating before filtering by
remote id.