refactor*: Remove `export!` macro
The `export!` macro not only introduces a new syntax for an otherwise
low-noise feature, it also prevents linter and formatter from working
correctly - hence this commit also removes a couple of files that are
actually empty, but weren't noticed by Clippy before.

Similarly, one file got mildly reformatted - that's because `cargo fmt`
used not to see that module before.