refactor(ET-3428)*: Add missing reply to field
We were missing the reply to from the message type. We also update
the type to store this in a separate table rather than an embedded
serialized json as it is easier to update in future revisions.

Since this is required for the next step, we need to make sure that all
existing messages and associated metadata is wiped from the db so we can
redownload all the missing metadata.