feat: Sanitize <style> from HTML
In order to stop stylesheet bleeding in the composer, we (both ET and
web team) have decided to sanitize <style>s that are provided by sender.

That means we will remove these from both `<head>` and `<body>` in both
composer and in decrypted message view.

It matches what Google is doing in the Gmail.