fix: Breaking change: Fix Rust handling of keychain errors
Errors/exceptions coming from native side (Swift/Kotlin) were causing
uniFFI to panic which led to a crash.
Fortunately 99.999% of cases keychain just works flawlesly so we never
had an occurrence of this case in the wild.

The crash came from derive macro of `uniffi::Error` that it cannot
handle lifting flat errors. Lifting in that case means "deserializing"
Swift/Kotlin type info Rust type.

This MR fixes the issue by removing flat_error attribute. It is not
needed in this case, however **it is a breaking change that requires
minor adjustment on mobile side**.